backend/requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
mcp==1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
supabase>=2.3.0
openai>=1.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
celery>=5.3.4
redis>=5.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
aiohttp>=3.9.1